requirements.txt
flask
torch
transformers
pycryptodome